fix: remove build hashes from conda env file
otherwise, it would be coupled to the cpu architecture.
For example, when I tried to recreate the conda env
on Github Actions, it said
that it could not find the
packages. That is because
I exported the environment
file from my Mac—which
has different checksums for Linux for each dependency.